refactoring for evm verifier
