+++ crates/integration/Cargo.toml
@@ -31,10 +31,6 @@ ff = "0.13"
 once_cell = "1.20"
 serde_json = "1.0"
 sha2 = "0.10"
-revm = "18.0"
-
-[dev-dependencies]
-snark-verifier-sdk = { git = "https://github.com/axiom-crypto/snark-verifier", branch = "zkvm-v0.1", default-features = false, features = ["loader_evm", "revm"] }
 
 [features]
 default = ["scroll"]
