--- crates/integration/src/utils/evm.rs
@@ -1,61 +1,0 @@
-use revm::{
-    Context, Evm, Handler, InMemoryDB,
-    primitives::{ExecutionResult, Output, TransactTo, TxEnv, TxKind, specification::CancunSpec},
-};
-
-/// Deploy contract and then call with calldata.
-/// Returns gas_used of call to deployed contract if both transactions are successful.
-pub fn deploy_and_call(deployment_code: Vec<u8>, calldata: Vec<u8>) -> Result<u64, String> {
-    let mut evm = Evm::new(
-        Context::new_with_db(InMemoryDB::default()),
-        Handler::mainnet::<CancunSpec>(),
-    );
-
-    *evm.tx_mut() = TxEnv {
-        gas_limit: u64::MAX,
-        transact_to: TxKind::Create,
-        data: deployment_code.into(),
-        ..Default::default()
-    };
-
-    let result = evm.transact_commit().unwrap();
-    let contract = match result {
-        ExecutionResult::Success {
-            output: Output::Create(_, Some(contract)),
-            ..
-        } => contract,
-        ExecutionResult::Revert { gas_used, output } => {
-            return Err(format!(
-                "Contract deployment transaction reverts with gas_used {gas_used} and output {:#x}",
-                output
-            ));
-        }
-        ExecutionResult::Halt { reason, gas_used } => {
-            return Err(format!(
-                "Contract deployment transaction halts unexpectedly with gas_used {gas_used} and reason {:?}",
-                reason
-            ));
-        }
-        _ => unreachable!(),
-    };
-
-    *evm.tx_mut() = TxEnv {
-        gas_limit: u64::MAX,
-        transact_to: TransactTo::Call(contract),
-        data: calldata.into(),
-        ..Default::default()
-    };
-
-    let result = evm.transact_commit().unwrap();
-    match result {
-        ExecutionResult::Success { gas_used, .. } => Ok(gas_used),
-        ExecutionResult::Revert { gas_used, output } => Err(format!(
-            "Contract call transaction reverts with gas_used {gas_used} and output {:#x}",
-            output
-        )),
-        ExecutionResult::Halt { reason, gas_used } => Err(format!(
-            "Contract call transaction halts unexpectedly with gas_used {gas_used} and reason {:?}",
-            reason
-        )),
-    }
-}
