+++ crates/prover/Cargo.toml
@@ -7,6 +7,7 @@ version.workspace = true
 
 [dependencies]
 scroll-zkvm-circuit-input-types.workspace = true
+scroll-zkvm-verifier.workspace = true
 
 alloy-primitives = { workspace = true, default-features = false, features = ["std", "map-hashbrown", "map-fxhash", "rkyv"] }
 rkyv.workspace = true
@@ -25,9 +26,9 @@ once_cell = "1.20"
 serde = "1.0"
 serde_json = "1.0"
 serde_stacker = "0.1"
-snark-verifier-sdk = { git = "https://github.com/axiom-crypto/snark-verifier", branch = "zkvm-v0.1", default-features = false, features = ["loader_evm", "revm"] }
 thiserror = "2.0"
 toml = "0.8"
+revm = "18.0"
 
 [features]
 default = ["scroll"]
