+++ crates/verifier/src/lib.rs
@@ -2,6 +2,9 @@ pub fn add(left: u64, right: u64) -> u64 {
     left + right
 }
 
+mod evm;
+pub use evm::*;
+
 #[cfg(test)]
 mod tests {
     use super::*;
